Graf a Double

+++ src/prop/g12/common/Graf.java
@@ -3,14 +3,14 @@ package prop.g12.common;
 import java.util.ArrayList;
 
 public class Graf<T> {
-	Integer[][] adjacencia;
+	Double[][] adjacencia;
 	Integer numNodes;
 	ArrayList<T> nodes;
 	int c;
 	
 	public void crearGraf(int n) {
 		numNodes = n;
-		adjacencia = new Integer[numNodes][numNodes];
+		adjacencia = new Double[numNodes][numNodes];
 		nodes = new ArrayList<T>(numNodes);
 		c = 0;
 	}
@@ -20,7 +20,7 @@ public class Graf<T> {
 		++c;
 	}
 	
-	public void afegirAresta(Integer idNode1, Integer  idNode2, Integer pes) {
+	public void afegirAresta(Integer idNode1, Integer  idNode2, Double pes) {
 		if (pes >= 0) {
 			adjacencia[idNode1][idNode2] = pes;
 			adjacencia[idNode2][idNode1] = pes;
@@ -39,7 +39,7 @@ public class Graf<T> {
 		return nodes;
 	}
 	
-	public Integer getAresta(int idNode1, int idNode2) {
+	public Double getAresta(int idNode1, int idNode2) {
 		return adjacencia[idNode1][idNode2];
 	}
